Generate markdown file from literate haskell for OAuth2 article
The command to do this is:

pandoc --shift-heading-level-by=1
       --wrap=preserve -t commonmark_x
       -o index.md oauth2-openid-connect-haskell.lhs

Using "commonmark_x" for extended syntax allows footnote extensions
which are needed here.

The metadata isn't included and needs to be copied across. Also the
language for code blocks needs to be edited to "haskell" since the
pandoc output for this format isn't recognised by Hugo's processor.